packages/lavender-data/pyproject.toml
[tool.poetry]
name = "lavender-data"
version = "0.0.1"
description = "Lavender Data"
authors = [
  "Hanch Han <cndghks15@gmail.com>"
]
repository = "https://github.com/fal-ai/lavender-data"
documentation = "https://github.com/fal-ai/lavender-data"
license = "Apache-2.0"
readme = "README.md"
packages = [
  {include = "lavender_data", from = "."},
  {include = "openapi_lavender_data_rest", from = "./openapi-lavender-data-rest"},
]

[tool.poetry.dependencies]
python = ">=3.10,<4"
numpy = "=2.1.2"
pyarrow = "=19.0.1"
ujson = "^5.10.0"
fastapi = {version = "0.115.6", extras = ["standard"]}
sqlmodel = "^0.0.23"
redis = {extras = ["hiredis"], version = "^5.2.1"}
httpx = ">=0.20.0,<0.29.0"
attrs = ">=22.2.0"
python-dateutil = "^2.8.0"
psycopg2-binary = "^2.9.10"
pydantic-settings = "^2.8.1"
openapi-lavender-data-rest = {path = "./openapi-lavender-data-rest"}

[tool.poetry.group.dev.dependencies]
autopep8 = "2.1.0"
flake8 = "7.0.0"
pytest = "8.1.1"
pytest-sugar = "1.0.0"
black = "^24.10.0"
openapi-python-client = "^0.24.0"
tqdm = "^4.67.1"

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"
